@@ -1,8 +1,7 @@
-import Vue from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
 
-Vue.config.productionTip = false
+import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app');
